adapt tests to fragments
